backend/requirements.txt
# FastAPI and server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
supabase==2.5.1
python-dotenv==1.0.1

# LangChain for agent orchestration
langchain==0.2.6
langchain-openai==0.1.14
langchain-anthropic==0.1.19
langchain-community==0.2.6

# Real-time
python-socketio==5.11.3

# Utilities
pydantic==2.8.2
pydantic-settings==2.3.4
httpx==0.27.0
